Updates 30-second walk instructions, naming
Updates 30-second walk instructions with the purple background color present in mPower to provide a consistent feel in advance of a refactoring to allow primary/secondary color to be set.  Also changes the naming conventions. Notably does *NOT* update the walk instructions on the WalkAndBalance task until able to speak with mPower2 team